Adicionado mod_shared. Removido inserted_by e updated_by do parametro das funções. Renomeado modelos do namespace para evitar confusões.
