+++ app/mod_auth/models.py
@@ -1,39 +1,35 @@
 # Import the database object (db) from the main application module
 from app import db
-
+from app.mod_shared.models import Base
 
 # Define a base model for other database tables to inherit
-class Base(db.Model):
+from app.mod_events.models import user_interest
+from app.mod_events.models import event_participation
 
-    __abstract__ = True
-
-    id = db.Column(db.Integer, primary_key=True)
-    date_created = db.Column(db.DateTime,  default=db.func.current_timestamp())
-    date_modified = db.Column(db.DateTime,  default=db.func.current_timestamp(),
-                              onupdate=db.func.current_timestamp())
-
-
-# Define a User model
 class User(Base):
+    id_user = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(80))
+    email = db.Column(db.String(50), unique=True)
+    sigaa_registration_number = db.Column(db.String(15), unique=True)
+    sigaa_user_name = db.Column(db.String(50), unique=True)
+    password = db.Column(db.String(64))
+    id_photo_file = db.Column(db.Integer, db.ForeignKey('file.id_file'))
+    event_participations = db.relationship(
+        'Event',
+        secondary=event_participation,
+        backref=db.backref('participating_users', lazy='dynamic')
+    )
+    user_interests = db.relationship(
+        'Tag',
+        secondary=user_interest,
+        backref=db.backref('interested_users', lazy='dynamic')
+    )
 
-    __tablename__ = 'auth_user'
-
-    # User Name
-    name = db.Column(db.String(128),  nullable=False)
-
-    # Identification Data: email & password
-    email = db.Column(db.String(128), nullable=False, unique=True)
-    password = db.Column(db.String(192), nullable=False)
-
-    # Authorisation Data: role & status
-    role = db.Column(db.SmallInteger, nullable=False)
-    status = db.Column(db.SmallInteger, nullable=False)
-
-    # New instance instantiation procedure
-    def __init__(self, name, email, password):
+    def __init__(self, name, email, sigaa_registration_number, sigaa_user_name,
+                 password, id_photo_file):
         self.name = name
         self.email = email
+        self.sigaa_registration_number = sigaa_registration_number
+        self.sigaa_user_name = sigaa_user_name
         self.password = password
-
-    def __repr__(self):
-        return '<User %r>' % self.name
+        self.id_photo_file = id_photo_file
